feat(FileService): integrate UploadService and refactor create, deleteById and deleteByOwner id handle files uploading

+++ backend/src/services/files/FileService.ts
@@ -1,18 +1,35 @@
 import { FileRepository } from '../../repositories/FileRepository';
 import { WithId } from '../../types/WithId';
 import { IFileService } from './IFileService';
-import { CreateFileDTO } from '../../dtos/caseFile/CreateFileDTO';
 import { FileDTO } from '../../dtos/caseFile/FileDTO';
 import { validateFile } from '../validators/files/validateFile';
 import { Page } from '../../types/Page';
 import { PageParams } from '../../types/PageParams';
 import { DeleteManyResult } from '../../types/DeleteManyResult';
+import { UploadService } from '../uṕload/UploadService';
+import { getFormatedFileName } from '../../utils/getFormatedFileName';
 
 export class FileService implements IFileService {
-  constructor(private fileRepository: FileRepository) {}
-  async create(data: CreateFileDTO, buffer: Buffer): Promise<WithId<FileDTO>> {
-    await validateFile(data, buffer);
-    return await this.fileRepository.create(data);
+  constructor(
+    private fileRepository: FileRepository,
+    private uploadService: UploadService
+  ) {}
+  async create(userId: string, ownerId: string, file: any): Promise<WithId<FileDTO>> {
+    await validateFile(file);
+    const uploadResult = await this.uploadService.upload(file.buffer);
+
+    const fileName = getFormatedFileName(file.originalname);
+
+    return await this.fileRepository.create({
+      ownerId,
+      name: fileName,
+      url: uploadResult.url,
+      downloadUrl: uploadResult.downloadUrl,
+      publicId: uploadResult.publicId,
+      size: file.size,
+      mimeType: file.mimetype,
+      uploadedBy: String(userId),
+    });
   }
 
   async findById(fileId: string): Promise<WithId<FileDTO>> {
@@ -32,10 +49,25 @@ export class FileService implements IFileService {
   }
 
   async deleteById(fileId: string): Promise<WithId<FileDTO> | null> {
+    const file = await this.findById(fileId);
+    if (!file) {
+      return null;
+    }
+    await this.uploadService.delete(file.publicId);
     return await this.fileRepository.deleteById(fileId);
   }
 
   async deleteByOwnerId(ownerId: string): Promise<DeleteManyResult> {
+    const files = await this.findAllByOwnerId(ownerId);
+    const filesPublicIds = files.map((file) => file.publicId);
+
+    const deleteUploadedResult = await this.uploadService.deleteMany(filesPublicIds);
+    if (deleteUploadedResult.failedCount > 0) {
+      console.log(
+        `Warning: ${deleteUploadedResult.failedCount} files could not be deleted from the storage`
+      );
+    }
+
     return await this.fileRepository.deleteByOwnerId(ownerId);
   }
 }
